+++ rollup.config.js
@@ -1,9 +1,17 @@
+import sourcemaps from 'rollup-plugin-sourcemaps';
+
 export const globals = {
   // Apollo
   'apollo-client': 'apollo.core',
   'apollo-link': 'apolloLink.core',
   'apollo-link-batch': 'apolloLink.batch',
   'apollo-utilities': 'apollo.utilities',
+
+  //GraphQL
+  'graphql/language/printer': 'printer',
+
+  //zen-observable
+  'zen-observable': 'Observable',
 };
 
 export default name => ({
@@ -18,6 +26,7 @@ export default name => ({
   },
   external: Object.keys(globals),
   onwarn,
+  plugins: [sourcemaps()],
 });
 
 export function onwarn(message) {